Added safer access and others
Changed MemCheck to be a class. Members are now private and can only be accessed with getter functions. Added default constructor for MemCheck. Added the ability to instantiate and utilize your own MemChecks. Added macro to define maximum number of MemCheck objects that can be instantiated in 1 program. Added the ability to enable and disable select MemChecks. Changed name of global MemCheck to be more readable. Changed end-program display toggle to be independent of MemCheck class. Added more examples at the bottom.